hedlin now working

+++ replicate/hedlin/utils/ptp_utils.py
@@ -19,10 +19,8 @@ from tqdm.notebook import tqdm
 import torch.nn.functional as F
 
 
-
 def diffusion_step(model, controller, latents, context, t, guidance_scale=None, cfg = True):
-    latents = latents.type(model.unet.dtype)
-    context = context.type(model.unet.dtype)
+    
     if cfg:
         latents_input = torch.cat([latents] * 2)
         noise_pred = model.unet(latents_input, t, encoder_hidden_states=context)["sample"]
